+++ app/models/draw.rb
@@ -24,7 +24,10 @@
 class Draw < ApplicationRecord # rubocop:disable ClassLength
   has_many :groups, dependent: :destroy
   has_many :draw_memberships, dependent: :destroy
-  has_many :students, through: :draw_memberships, source: :user
+  has_many :students,
+           -> { joins(:draws).where('draws.active = draw_memberships.active') },
+           through: :draw_memberships, source: :user
+
   has_many :draw_suites, dependent: :delete_all
   has_many :suites, through: :draw_suites
   has_many :lottery_assignments, dependent: :destroy